scripts/outreach/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
python-dotenv>=1.0.0
pandas>=2.0.0
markdown>=3.5.0
jinja2>=3.1.0
email-validator>=2.1.0
schedule>=1.2.0
lxml>=4.9.0
